Separate worker into another file
