+++ packages/component/.babelrc
@@ -1,11 +1,25 @@
 {
   "env": {
     "test": {
+      "exclude": [
+        "src/Utils/downscaleImageToDataURL.worker.js"
+      ],
       "plugins": [
         "babel-plugin-istanbul"
       ]
     }
   },
+  "overrides": [{
+    "test": "src/Utils/downscaleImageToDataURL.worker.js",
+    "plugins": [],
+    "presets": [
+      ["@babel/preset-env", {
+        "targets": {
+          "chrome": "69"
+        }
+      }]
+    ]
+  }],
   "plugins": [
     "@babel/proposal-object-rest-spread",
     [
